+++ packages/webview_flutter/webview_flutter_wkwebview/test/webkit_webview_cookie_manager_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in webview_flutter_wkwebview/example/ios/.symlinks/plugins/webview_flutter_wkwebview/test/v4/webkit_webview_widget_test.dart.
+// in webview_flutter_wkwebview/example/ios/.symlinks/plugins/webview_flutter_wkwebview/test/webkit_webview_cookie_manager_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -21,9 +21,9 @@ import 'package:webview_flutter_wkwebview/src/web_kit/web_kit.dart' as _i2;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeWKUserContentController_0 extends _i1.SmartFake
-    implements _i2.WKUserContentController {
-  _FakeWKUserContentController_0(
+class _FakeWKHttpCookieStore_0 extends _i1.SmartFake
+    implements _i2.WKHttpCookieStore {
+  _FakeWKHttpCookieStore_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -32,19 +32,9 @@ class _FakeWKUserContentController_0 extends _i1.SmartFake
         );
 }
 
-class _FakeWKPreferences_1 extends _i1.SmartFake implements _i2.WKPreferences {
-  _FakeWKPreferences_1(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-class _FakeWKWebsiteDataStore_2 extends _i1.SmartFake
+class _FakeWKWebsiteDataStore_1 extends _i1.SmartFake
     implements _i2.WKWebsiteDataStore {
-  _FakeWKWebsiteDataStore_2(
+  _FakeWKWebsiteDataStore_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -53,86 +43,119 @@ class _FakeWKWebsiteDataStore_2 extends _i1.SmartFake
         );
 }
 
-class _FakeWKWebViewConfiguration_3 extends _i1.SmartFake
-    implements _i2.WKWebViewConfiguration {
-  _FakeWKWebViewConfiguration_3(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-/// A class which mocks [WKWebViewConfiguration].
+/// A class which mocks [WKWebsiteDataStore].
 ///
 /// See the documentation for Mockito's code generation for more information.
 // ignore: must_be_immutable
-class MockWKWebViewConfiguration extends _i1.Mock
-    implements _i2.WKWebViewConfiguration {
-  MockWKWebViewConfiguration() {
+class MockWKWebsiteDataStore extends _i1.Mock
+    implements _i2.WKWebsiteDataStore {
+  MockWKWebsiteDataStore() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.WKUserContentController get userContentController => (super.noSuchMethod(
-        Invocation.getter(#userContentController),
-        returnValue: _FakeWKUserContentController_0(
+  _i2.WKHttpCookieStore get httpCookieStore => (super.noSuchMethod(
+        Invocation.getter(#httpCookieStore),
+        returnValue: _FakeWKHttpCookieStore_0(
           this,
-          Invocation.getter(#userContentController),
+          Invocation.getter(#httpCookieStore),
         ),
-      ) as _i2.WKUserContentController);
+      ) as _i2.WKHttpCookieStore);
   @override
-  _i2.WKPreferences get preferences => (super.noSuchMethod(
-        Invocation.getter(#preferences),
-        returnValue: _FakeWKPreferences_1(
-          this,
-          Invocation.getter(#preferences),
-        ),
-      ) as _i2.WKPreferences);
-  @override
-  _i2.WKWebsiteDataStore get websiteDataStore => (super.noSuchMethod(
-        Invocation.getter(#websiteDataStore),
-        returnValue: _FakeWKWebsiteDataStore_2(
-          this,
-          Invocation.getter(#websiteDataStore),
-        ),
-      ) as _i2.WKWebsiteDataStore);
-  @override
-  _i3.Future<void> setAllowsInlineMediaPlayback(bool? allow) =>
+  _i3.Future<bool> removeDataOfTypes(
+    Set<_i2.WKWebsiteDataType>? dataTypes,
+    DateTime? since,
+  ) =>
       (super.noSuchMethod(
         Invocation.method(
-          #setAllowsInlineMediaPlayback,
-          [allow],
+          #removeDataOfTypes,
+          [
+            dataTypes,
+            since,
+          ],
         ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
+        returnValue: _i3.Future<bool>.value(false),
+      ) as _i3.Future<bool>);
   @override
-  _i3.Future<void> setMediaTypesRequiringUserActionForPlayback(
-          Set<_i2.WKAudiovisualMediaType>? types) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #setMediaTypesRequiringUserActionForPlayback,
-          [types],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-  @override
-  _i2.WKWebViewConfiguration copy() => (super.noSuchMethod(
+  _i2.WKWebsiteDataStore copy() => (super.noSuchMethod(
         Invocation.method(
           #copy,
           [],
         ),
-        returnValue: _FakeWKWebViewConfiguration_3(
+        returnValue: _FakeWKWebsiteDataStore_1(
           this,
           Invocation.method(
             #copy,
             [],
           ),
         ),
-      ) as _i2.WKWebViewConfiguration);
+      ) as _i2.WKWebsiteDataStore);
+  @override
+  _i3.Future<void> addObserver(
+    _i4.NSObject? observer, {
+    required String? keyPath,
+    required Set<_i4.NSKeyValueObservingOptions>? options,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #addObserver,
+          [observer],
+          {
+            #keyPath: keyPath,
+            #options: options,
+          },
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
+  @override
+  _i3.Future<void> removeObserver(
+    _i4.NSObject? observer, {
+    required String? keyPath,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeObserver,
+          [observer],
+          {#keyPath: keyPath},
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
+}
+
+/// A class which mocks [WKHttpCookieStore].
+///
+/// See the documentation for Mockito's code generation for more information.
+// ignore: must_be_immutable
+class MockWKHttpCookieStore extends _i1.Mock implements _i2.WKHttpCookieStore {
+  MockWKHttpCookieStore() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i3.Future<void> setCookie(_i4.NSHttpCookie? cookie) => (super.noSuchMethod(
+        Invocation.method(
+          #setCookie,
+          [cookie],
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
+  @override
+  _i2.WKHttpCookieStore copy() => (super.noSuchMethod(
+        Invocation.method(
+          #copy,
+          [],
+        ),
+        returnValue: _FakeWKHttpCookieStore_0(
+          this,
+          Invocation.method(
+            #copy,
+            [],
+          ),
+        ),
+      ) as _i2.WKHttpCookieStore);
   @override
   _i3.Future<void> addObserver(
     _i4.NSObject? observer, {
